# docs/openapi/hansapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/data-go-kr/hira/hospitals/{ykiho}": {
      "get": {
        "description": "암호화된 요양기호로 1건 조회. 없으면 items 가 빈 배열이고 totalCount 가 0 이다.",
        "operationId": "HiraHospitalController_get",
        "parameters": [
          {
            "name": "ykiho",
            "required": true,
            "in": "path",
            "description": "암호화된 요양기호. 목록 응답의 ykiho 를 그대로 쓴다.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HiraHospitalListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "병원 상세 조회",
        "tags": [
          "hira"
        ]
      }
    },
...
    "/data-go-kr/nmc/hospitals": {
      "get": {
        "description": "로컬 DB 에 미러링한 NMC 병원 목록. 응답 구조는 원본 API 와 동일하다.",
        "operationId": "NmcHospitalController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmcHospitalFullDownResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "병원 목록 조회",
        "tags": [
          "nmc"
        ]
      }
    },
    "/data-go-kr/nmc/hospitals/{hpid}": {
      "get": {
        "description": "기관ID 로 1건 조회. 없으면 items 가 빈 배열이고 totalCount 가 0 이다(원본 API 와 동일).",
        "operationId": "NmcHospitalController_get",
        "parameters": [
          {
            "name": "hpid",
            "required": true,
            "in": "path",
            "description": "기관ID",
            "schema": {
              "example": "A1118361",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmcHospitalFullDownResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "병원 상세 조회",
        "tags": [
          "nmc"
        ]
      }
    },
...
    "/data-go-kr/nmc/regions": {
      "get": {
        "description": "병원 데이터에서 집계한 시도·시군구 목록. 병원이 1건 이상 있는 지역만 나온다.",
        "operationId": "NmcRegionController_listRegions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmcRegionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "지역 목록 조회",
        "tags": [
          "nmc"
        ]
      }
    },
    "/data-go-kr/nmc/hospitals/{hpid}/subjects": {
      "get": {
        "description": "해당 병원이 진료하는 과목 목록. 코드마스터 D000 기준이다.",
        "operationId": "NmcRegionController_listSubjects",
        "parameters": [
          {
            "name": "hpid",
            "required": true,
            "in": "path",
            "description": "기관ID",
            "schema": {
              "example": "A1100001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmcSubjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "병원 진료과목 조회",
        "tags": [
          "nmc"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hans API",
    "description": "Hans API backend 문서",
    "version": "0.15.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local"
    },
    {
      "url": "https://develop-api.plzhans.com",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT or service key(sk_...)",
        "type": "http",
        "description": "사용자 access token(JWT) 또는 서비스 키(sk_{appId}_{keyId}_{rand})."
      },
      "client-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "공개 클라이언트 식별자(clientId). WEB 은 Origin 도 검사됨."
      }
    },
    "schemas": {
...
      "KrDataHeaderDto": {
        "type": "object",
        "properties": {
          "resultCode": {
            "type": "string",
            "description": "결과 코드. 정상은 00",
            "example": "00"
          },
          "resultMsg": {
            "type": "string",
            "description": "결과 메시지",
            "example": "NORMAL SERVICE."
          }
        },
        "required": [
          "resultCode",
          "resultMsg"
        ]
      },
...
      "NmcRegionItemDto": {
        "type": "object",
        "properties": {
          "sidoNm": {
            "type": "string",
            "description": "시도명",
            "example": "서울특별시"
          },
          "sgguNm": {
            "type": "string",
            "description": "시군구명. 세종특별자치시처럼 시군구가 없는 지역은 값이 없다.",
            "example": "강남구"
          },
          "hospitalCount": {
            "type": "number",
            "description": "이 지역의 병원 수",
            "example": 3045
          }
        },
        "required": [
          "sidoNm",
          "hospitalCount"
        ]
      },
      "NmcRegionItemsDto": {
        "type": "object",
        "properties": {
          "item": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NmcRegionItemDto"
            }
          }
        },
        "required": [
          "item"
        ]
      },
      "NmcRegionBodyDto": {
        "type": "object",
        "properties": {
          "items": {
            "$ref": "#/components/schemas/NmcRegionItemsDto"
          },
          "numOfRows": {
            "type": "number",
            "description": "한 페이지 결과 수",
            "example": 100
          },
          "pageNo": {
            "type": "number",
            "description": "페이지 번호",
            "example": 1
          },
          "totalCount": {
            "type": "number",
            "description": "전체 결과 수",
            "example": 263
          }
        },
        "required": [
          "items",
          "numOfRows",
          "pageNo",
          "totalCount"
        ]
      },
      "NmcRegionResultDto": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/KrDataHeaderDto"
          },
          "body": {
            "$ref": "#/components/schemas/NmcRegionBodyDto"
          }
        },
        "required": [
          "header",
          "body"
        ]
      },
      "NmcRegionResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/NmcRegionResultDto"
          }
        },
        "required": [
          "response"
        ]
      },
      "NmcSubjectItemDto": {
        "type": "object",
        "properties": {
          "subjectCd": {
            "type": "string",
            "description": "진료과목 코드 (코드마스터 D000)",
            "example": "D001"
          },
          "subjectNm": {
            "type": "string",
            "description": "진료과목명",
            "example": "내과"
          }
        },
        "required": [
          "subjectCd"
        ]
      },
      "NmcSubjectItemsDto": {
        "type": "object",
        "properties": {
          "item": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NmcSubjectItemDto"
            }
          }
        },
        "required": [
          "item"
        ]
      },
      "NmcSubjectBodyDto": {
        "type": "object",
        "properties": {
          "items": {
            "$ref": "#/components/schemas/NmcSubjectItemsDto"
          },
          "numOfRows": {
            "type": "number",
            "description": "한 페이지 결과 수",
            "example": 25
          },
          "pageNo": {
            "type": "number",
            "description": "페이지 번호",
            "example": 1
          },
          "totalCount": {
            "type": "number",
            "description": "전체 결과 수",
            "example": 25
          }
        },
        "required": [
          "items",
          "numOfRows",
          "pageNo",
          "totalCount"
        ]
      },
      "NmcSubjectResultDto": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/KrDataHeaderDto"
          },
          "body": {
            "$ref": "#/components/schemas/NmcSubjectBodyDto"
          }
        },
        "required": [
          "header",
          "body"
        ]
      },
      "NmcSubjectResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/NmcSubjectResultDto"
          }
        },
        "required": [
          "response"
        ]
      },
...
      "NmcResultHeader": {
        "type": "object",
        "description": "공통 응답 헤더",
        "properties": {
          "resultCode": {
            "type": "string",
            "description": "결과코드 (정상: 00)"
          },
          "resultMsg": {
            "type": "string",
            "description": "결과메시지"
          }
        }
      },
      "NmcPageInfo": {
        "type": "object",
        "properties": {
          "numOfRows": {
            "type": "integer",
            "description": "한 페이지 결과 수"
          },
          "pageNo": {
            "type": "integer",
            "description": "현재 페이지 번호"
          },
          "totalCount": {
            "type": "integer",
            "description": "전체 결과 수"
          }
        }
      },
...
      "NmcHospitalFullDownResponse": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/NmcHospitalFullDownResult"
          }
        }
      },
...
      "NmcHospitalFullDownItem": {
        "type": "object",
        "description": "병/의원 FullData 항목",
        "properties": {
          "rnum": {
            "type": "integer",
            "description": "일련번호"
          },
          "hpid": {
            "type": "string",
            "description": "기관ID"
          },
          "dutyName": {
            "type": "string",
            "description": "기관명"
          },
          "dutyAddr": {
            "type": "string",
            "description": "주소"
          },
          "dutyDiv": {
            "$ref": "#/components/schemas/NmcDutyDiv"
          },
          "dutyDivNam": {
            "type": "string",
            "description": "병원분류명"
          },
          "dutyEmcls": {
            "type": "string",
            "description": "응급의료기관코드"
          },
          "dutyEmclsName": {
            "type": "string",
            "description": "응급의료기관코드명"
          },
          "dutyEryn": {
            "type": "string",
            "description": "응급실운영여부 (1:운영, 2:미운영)"
          },
          "dutyEtc": {
            "type": "string",
            "description": "비고"
          },
          "dutyInf": {
            "type": "string",
            "description": "기관설명상세"
          },
          "dutyMapimg": {
            "type": "string",
            "description": "간이약도"
          },
          "dutyTel1": {
            "type": "string",
            "description": "대표전화1"
          },
          "dutyTel3": {
            "type": "string",
            "description": "응급실전화"
          },
          "postCdn1": {
            "type": "string",
            "description": "우편번호1"
          },
          "postCdn2": {
            "type": "string",
            "description": "우편번호2"
          },
          "wgs84Lon": {
            "type": "string",
            "description": "병원경도"
          },
          "wgs84Lat": {
            "type": "string",
            "description": "병원위도"
          },
          "dutyTime1s": {
            "type": "string",
            "description": "진료시간(월요일) 시작"
          },
          "dutyTime2s": {
            "type": "string",
            "description": "진료시간(화요일) 시작"
          },
          "dutyTime3s": {
            "type": "string",
            "description": "진료시간(수요일) 시작"
          },
          "dutyTime4s": {
            "type": "string",
            "description": "진료시간(목요일) 시작"
          },
          "dutyTime5s": {
            "type": "string",
            "description": "진료시간(금요일) 시작"
          },
          "dutyTime6s": {
            "type": "string",
            "description": "진료시간(토요일) 시작"
          },
          "dutyTime7s": {
            "type": "string",
            "description": "진료시간(일요일) 시작"
          },
          "dutyTime8s": {
            "type": "string",
            "description": "진료시간(공휴일) 시작"
          },
          "dutyTime1c": {
            "type": "string",
            "description": "진료시간(월요일) 종료"
          },
          "dutyTime2c": {
            "type": "string",
            "description": "진료시간(화요일) 종료"
          },
          "dutyTime3c": {
            "type": "string",
            "description": "진료시간(수요일) 종료"
          },
          "dutyTime4c": {
            "type": "string",
            "description": "진료시간(목요일) 종료"
          },
          "dutyTime5c": {
            "type": "string",
            "description": "진료시간(금요일) 종료"
          },
          "dutyTime6c": {
            "type": "string",
            "description": "진료시간(토요일) 종료"
          },
          "dutyTime7c": {
            "type": "string",
            "description": "진료시간(일요일) 종료"
          },
          "dutyTime8c": {
            "type": "string",
            "description": "진료시간(공휴일) 종료"
          },
          "sidoNm": {
            "type": "string",
            "description": "시도명 (예: 서울특별시).\n\n**원본 API 에는 없는 파생 필드다.** 주소(dutyAddr)의 첫 토큰에서 추출하며,\n축약 표기(서울)는 정식 명칭(서울특별시)으로 정규화한다.",
            "example": "서울특별시"
          },
          "sgguNm": {
            "type": "string",
            "description": "시군구명 (예: 강남구).\n\n**원본 API 에는 없는 파생 필드다.** 주소(dutyAddr)의 둘째 토큰에서 추출한다.\n시 아래 구가 있으면 두 토큰을 합친다(예: 수원시 팔달구).\n세종특별자치시처럼 시군구가 없는 지역은 값이 없다.",
            "example": "강남구"
          }
        }
      },
...
      "NmcHospitalFullDownItems": {
        "type": "object",
        "properties": {
          "item": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NmcHospitalFullDownItem"
            }
          }
        }
      },
      "NmcHospitalFullDownBody": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NmcPageInfo"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/NmcHospitalFullDownItems"
              }
            }
          }
        ]
      },
      "NmcHospitalFullDownResult": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/NmcResultHeader"
          },
          "body": {
            "$ref": "#/components/schemas/NmcHospitalFullDownBody"
          }
        }
      },
...
      "HiraResultHeader": {
        "type": "object",
        "description": "공통 응답 헤더",
        "properties": {
          "resultCode": {
            "type": "string",
            "description": "결과코드 (정상: 00)"
          },
          "resultMsg": {
            "type": "string",
            "description": "결과메시지"
          }
        }
      },
      "HiraPageInfo": {
        "type": "object",
        "properties": {
          "numOfRows": {
            "type": "integer",
            "description": "한 페이지 결과 수"
          },
          "pageNo": {
            "type": "integer",
            "description": "현재 페이지 번호"
          },
          "totalCount": {
            "type": "integer",
            "description": "전체 결과 수"
          }
        }
      },
      "HiraHospitalListResponse": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/HiraHospitalListResult"
          }
        }
      },
      "HiraHospitalItem": {
        "type": "object",
        "description": "병원 기본정보 항목",
        "properties": {
          "ykiho": {
            "type": "string",
            "description": "암호화된 요양기호. 상세정보 조회(MadmDtlInfoService2.8)의 키다."
          },
          "yadmNm": {
            "type": "string",
            "description": "병원명"
          },
          "clCd": {
            "type": "string",
            "description": "종별코드"
          },
          "clCdNm": {
            "type": "string",
            "description": "종별코드명 (예: 상급종합)"
          },
          "sidoCd": {
            "type": "string",
            "description": "시도코드"
          },
          "sidoCdNm": {
            "type": "string",
            "description": "시도명"
          },
          "sgguCd": {
            "type": "string",
            "description": "시군구코드"
          },
          "sgguCdNm": {
            "type": "string",
            "description": "시군구명"
          },
          "emdongNm": {
            "type": "string",
            "description": "읍면동명"
          },
          "postNo": {
            "type": "string",
            "description": "우편번호"
          },
          "addr": {
            "type": "string",
            "description": "주소"
          },
          "telno": {
            "type": "string",
            "description": "전화번호"
          },
          "hospUrl": {
            "type": "string",
            "description": "홈페이지"
          },
          "estbDd": {
            "type": "string",
            "description": "개설일자 (YYYYMMDD)"
          },
          "drTotCnt": {
            "type": "integer",
            "description": "의사총수"
          },
          "mdeptGdrCnt": {
            "type": "integer",
            "description": "의과일반의 인원수"
          },
          "mdeptIntnCnt": {
            "type": "integer",
            "description": "의과인턴 인원수"
          },
          "mdeptResdntCnt": {
            "type": "integer",
            "description": "의과레지던트 인원수"
          },
          "mdeptSdrCnt": {
            "type": "integer",
            "description": "의과전문의 인원수"
          },
          "detyGdrCnt": {
            "type": "integer",
            "description": "치과일반의 인원수"
          },
          "detyIntnCnt": {
            "type": "integer",
            "description": "치과인턴 인원수"
          },
          "detyResdntCnt": {
            "type": "integer",
            "description": "치과레지던트 인원수"
          },
          "detySdrCnt": {
            "type": "integer",
            "description": "치과전문의 인원수"
          },
          "cmdcGdrCnt": {
            "type": "integer",
            "description": "한방일반의 인원수"
          },
          "cmdcIntnCnt": {
            "type": "integer",
            "description": "한방인턴 인원수"
          },
          "cmdcResdntCnt": {
            "type": "integer",
            "description": "한방레지던트 인원수"
          },
          "cmdcSdrCnt": {
            "type": "integer",
            "description": "한방전문의 인원수"
          },
          "pnursCnt": {
            "type": "integer",
            "description": "공식 가이드(2021)에 없는 필드. 값이 대부분 0 이고 드물게 한 자릿수인 것으로 보아\n전문간호사 수로 추정된다."
          },
          "XPos": {
            "type": "string",
            "description": "x좌표 (경도). 대문자 X 로 시작한다."
          },
          "YPos": {
            "type": "string",
            "description": "y좌표 (위도). 대문자 Y 로 시작한다."
          },
          "distance": {
            "type": "string",
            "description": "거리(m). xPos/yPos/radius 로 반경 검색했을 때만 온다."
          }
        }
      },
      "HiraHospitalListItems": {
        "type": "object",
        "properties": {
          "item": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HiraHospitalItem"
            }
          }
        }
      },
      "HiraHospitalListBody": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HiraPageInfo"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/HiraHospitalListItems"
              }
            }
          }
        ]
      },
      "HiraHospitalListResult": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/HiraResultHeader"
          },
          "body": {
            "$ref": "#/components/schemas/HiraHospitalListBody"
          }
        }
      },
...
    }
  }
}